Separate tests into multiple files inside a folder
