+++ test/async.js
@@ -1,0 +1,21 @@
+import test from "../src/test.js";
+
+const fakeAsyncTask = ({ duration = 30, returnValue = undefined } = {}) => {
+  return new Promise(resolve => setTimeout(resolve, duration, returnValue));
+};
+
+test("Supports async", async t => {
+  const result = await fakeAsyncTask({ returnValue: true });
+  t.assert(result);
+});
+
+test("Throwing inside async tests fails the test", async t => {
+  await fakeAsyncTask();
+  throw Error();
+});
+
+const slowest = 60;
+test(`Runs tests concurrently, all finish in about ${slowest}ms`, async t => {
+  await fakeAsyncTask({ duration: slowest });
+  t.assert(true);
+});
